backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pydantic==2.9.2
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
python-multipart==0.0.12
python-dotenv==1.0.1
email-validator==2.3.0
openai>=1.0.0
numpy>=1.24.0
psycopg2-binary>=2.9.9
pgvector>=0.3.0

